refactor(contented-processor): move markdown to its directory to separate concerns
