+++ packages/contented-processor/src/index.ts
@@ -1,4 +1,4 @@
 export * from './ContentedCodegen.js';
 export * from './ContentedPipeline.js';
 export * from './ContentedProcessor.js';
-export * from './ContentedUnified.js';
+export * from './markdown/UnifiedProcessor.js';
